backend_API/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.9
pillow==10.2.0